use isize as type to specify a core

+++ src/libstd/sys/hermit/thread.rs
@@ -35,7 +35,7 @@ pub const NORMAL_PRIO: Priority = Priority::from(2);
 
 extern "C" {
     fn sys_usleep(usecs: u64);
-    fn sys_spawn(id: *mut Tid, func: extern "C" fn(usize), arg: usize, prio: u8, core_id: i32) -> i32;
+    fn sys_spawn(id: *mut Tid, func: extern "C" fn(usize), arg: usize, prio: u8, core_id: isize) -> i32;
     fn sys_join(id: Tid) -> i32;
     fn sys_yield();
 }
@@ -50,7 +50,7 @@ unsafe impl Sync for Thread {}
 pub const DEFAULT_MIN_STACK_SIZE: usize = 262144;
 
 impl Thread {
-    pub unsafe fn new_with_coreid(_stack: usize, p: Box<dyn FnOnce()>, core_id: i32)
+    pub unsafe fn new_with_coreid(_stack: usize, p: Box<dyn FnOnce()>, core_id: isize)
         -> io::Result<Thread>
     {
         let p = box p;
